Struct near_primitives_core::config::ExtCostsConfig

pub struct ExtCostsConfig {

    pub base: Gas,
    pub contract_compile_base: Gas,
    pub contract_compile_bytes: Gas,
    pub read_memory_base: Gas,
    pub read_memory_byte: Gas,
    pub write_memory_base: Gas,
    pub write_memory_byte: Gas,
    pub read_register_base: Gas,
    pub read_register_byte: Gas,
    pub write_register_base: Gas,
    pub write_register_byte: Gas,
    pub utf8_decoding_base: Gas,
    pub utf8_decoding_byte: Gas,
    pub utf16_decoding_base: Gas,
    pub utf16_decoding_byte: Gas,
    pub sha256_base: Gas,
    pub sha256_byte: Gas,
    pub keccak256_base: Gas,
    pub keccak256_byte: Gas,
    pub keccak512_base: Gas,
    pub keccak512_byte: Gas,
    pub log_base: Gas,
    pub log_byte: Gas,
    pub storage_write_base: Gas,
    pub storage_write_key_byte: Gas,
    pub storage_write_value_byte: Gas,
    pub storage_write_evicted_byte: Gas,
    pub storage_read_base: Gas,
    pub storage_read_key_byte: Gas,
    pub storage_read_value_byte: Gas,
    pub storage_remove_base: Gas,
    pub storage_remove_key_byte: Gas,
    pub storage_remove_ret_value_byte: Gas,
    pub storage_has_key_base: Gas,
    pub storage_has_key_byte: Gas,
    pub storage_iter_create_prefix_base: Gas,
    pub storage_iter_create_prefix_byte: Gas,
    pub storage_iter_create_range_base: Gas,
    pub storage_iter_create_from_byte: Gas,
    pub storage_iter_create_to_byte: Gas,
    pub storage_iter_next_base: Gas,
    pub storage_iter_next_key_byte: Gas,
    pub storage_iter_next_value_byte: Gas,
    pub touching_trie_node: Gas,
    pub promise_and_base: Gas,
    pub promise_and_per_promise: Gas,
    pub promise_return: Gas,
    pub validator_stake_base: Gas,
    pub validator_total_stake_base: Gas,

}

Fields
base: Gas

Base cost for calling a host function.

contract_compile_base: Gas

Base cost of loading and compiling contract

contract_compile_bytes: Gas

Cost of the execution to load and compile contract

read_memory_base: Gas

Base cost for guest memory read

read_memory_byte: Gas

Cost for guest memory read

write_memory_base: Gas

Base cost for guest memory write

write_memory_byte: Gas

Cost for guest memory write per byte

read_register_base: Gas

Base cost for reading from register

read_register_byte: Gas

Cost for reading byte from register

write_register_base: Gas

Base cost for writing into register

write_register_byte: Gas

Cost for writing byte into register

utf8_decoding_base: Gas

Base cost of decoding utf8. It’s used for log_utf8 and panic_utf8.

utf8_decoding_byte: Gas

Cost per byte of decoding utf8. It’s used for log_utf8 and panic_utf8.

utf16_decoding_base: Gas

Base cost of decoding utf16. It’s used for log_utf16.

utf16_decoding_byte: Gas

Cost per byte of decoding utf16. It’s used for log_utf16.

sha256_base: Gas

Cost of getting sha256 base

sha256_byte: Gas

Cost of getting sha256 per byte

keccak256_base: Gas

Cost of getting sha256 base

keccak256_byte: Gas

Cost of getting sha256 per byte

keccak512_base: Gas

Cost of getting sha256 base

keccak512_byte: Gas

Cost of getting sha256 per byte

log_base: Gas

Cost for calling logging.

log_byte: Gas

Cost for logging per byte

storage_write_base: Gas

Storage trie write key base cost

storage_write_key_byte: Gas

Storage trie write key per byte cost

storage_write_value_byte: Gas

Storage trie write value per byte cost

storage_write_evicted_byte: Gas

Storage trie write cost per byte of evicted value.

storage_read_base: Gas

Storage trie read key base cost

storage_read_key_byte: Gas

Storage trie read key per byte cost

storage_read_value_byte: Gas

Storage trie read value cost per byte cost

storage_remove_base: Gas

Remove key from trie base cost

storage_remove_key_byte: Gas

Remove key from trie per byte cost

storage_remove_ret_value_byte: Gas

Remove key from trie ret value byte cost

storage_has_key_base: Gas

Storage trie check for key existence cost base

storage_has_key_byte: Gas

Storage trie check for key existence per key byte

storage_iter_create_prefix_base: Gas

Create trie prefix iterator cost base

storage_iter_create_prefix_byte: Gas

Create trie prefix iterator cost per byte.

storage_iter_create_range_base: Gas

Create trie range iterator cost base

storage_iter_create_from_byte: Gas

Create trie range iterator cost per byte of from key.

storage_iter_create_to_byte: Gas

Create trie range iterator cost per byte of to key.

storage_iter_next_base: Gas

Trie iterator per key base cost

storage_iter_next_key_byte: Gas

Trie iterator next key byte cost

storage_iter_next_value_byte: Gas

Trie iterator next key byte cost

touching_trie_node: Gas

Cost per touched trie node

promise_and_base: Gas

Cost for calling promise_and

promise_and_per_promise: Gas

Cost for calling promise_and for each promise

promise_return: Gas

Cost for calling promise_return

validator_stake_base: Gas

Cost of calling validator_stake.

validator_total_stake_base: Gas

Cost of calling validator_total_stake.