Module near_primitives_core::runtime::fees

Describes the various costs incurred by creating receipts. We use the following abbreviation for readability:

sir – sender is receiver. Receipts that are directed by an account to itself are guaranteed to not be cross-shard which is cheaper than cross-shard. Conversely, when sender is not a receiver it might or might not be a cross-shard communication.

Modules

u128_dec_format
Serde serializer for u128 to integer. This is copy from core/primitives/src/serialize.rs It is required as this module doesn’t depend on primitives. TODO(3384): move basic primitives into a separate module and use in runtime.

Structs

AccessKeyCreationConfig
Describes the cost of creating an access key.

ActionCreationConfig
Describes the cost of creating a specific action, Action. Includes all variants.

DataReceiptCreationConfig
Describes the cost of creating a data receipt, DataReceipt.

EvmBls12ConstOpCost

EvmBn128PairingCost

EvmCostConfig
Describe cost of evm

EvmLinearCost

EvmModexpCost

EvmPrecompileCostConfig

Fee
Costs associated with an object that can only be sent over the network (and executed by the receiver). NOTE: send_sir or send_not_sir fees are usually burned when the item is being created. And execution fee is burned when the item is being executed.

RuntimeFeesConfig

StorageUsageConfig
Describes cost of storage per block

Type Definitions

EvmBlake2FCost